Move network specific tasks to under network
