+++ packages/lodestar/src/node/nodejs.ts
@@ -126,7 +126,6 @@ export class BeaconNode {
       sync: this.sync,
       network: this.network,
       logger: this.logger.child(this.conf.logger.chores),
-      reps: this.reps,
     });
   }
