+++ packages/lodestar/src/sync/interface.ts
@@ -3,7 +3,6 @@ import {INetwork} from "../network";
 import {IReputationStore} from "./IReputation";
 import {ILogger} from "@chainsafe/lodestar-utils/lib/logger";
 import {CommitteeIndex, Slot, SyncingStatus} from "@chainsafe/lodestar-types";
-import StrictEventEmitter from "strict-event-emitter-types";
 import {InitialSync} from "./initial";
 import {IRegularSync} from "./regular";
 import {IGossipHandler} from "./gossip";
@@ -12,19 +11,12 @@ import {IBeaconChain} from "../chain";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
 import {IBeaconDb} from "../db/api";
 import {AttestationCollector} from "./utils";
-import {EventEmitter} from "events";
 
-export interface IBeaconSync extends IService, SyncEventEmitter {
+export interface IBeaconSync extends IService {
   getSyncStatus(): Promise<SyncingStatus>;
   isSynced(): boolean;
   collectAttestations(slot: Slot, committeeIndex: CommitteeIndex): Promise<void>;
 }
-
-export interface ISyncEvents {
-  syncCompleted: () => void;
-}
-
-export type SyncEventEmitter = StrictEventEmitter<EventEmitter, ISyncEvents>;
 
 export interface ISyncModule {
   getHighestBlock(): Slot;
