+++ packages/lodestar/src/sync/regular/interface.ts
@@ -4,7 +4,7 @@ import {EventEmitter} from "events";
 import StrictEventEmitter from "strict-event-emitter-types";
 
 export interface IRegularSyncEvents {
-  regularSyncCompleted: () => void;
+  syncCompleted: () => void;
 }
 
 export type RegularSyncEventEmitter = StrictEventEmitter<EventEmitter, IRegularSyncEvents>;
