+++ packages/lodestar/src/sync/sync.ts
@@ -1,5 +1,5 @@
 import PeerId from "peer-id";
-import {IBeaconSync, ISyncModules, SyncEventEmitter} from "./interface";
+import {IBeaconSync, ISyncModules} from "./interface";
 import defaultOptions, {ISyncOptions} from "./options";
 import {getSyncProtocols, INetwork} from "../network";
 import {IReputationStore} from "./IReputation";
@@ -15,7 +15,6 @@ import {IBeaconChain} from "../chain";
 import {NaiveRegularSync} from "./regular/naive";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
 import {List} from "@chainsafe/ssz";
-import {EventEmitter} from "events";
 
 export enum SyncMode {
   WAITING_PEERS,
@@ -25,7 +24,7 @@ export enum SyncMode {
   STOPPED,
 }
 
-export class BeaconSync extends (EventEmitter as {new (): SyncEventEmitter}) implements IBeaconSync {
+export class BeaconSync implements IBeaconSync {
   private readonly opts: ISyncOptions;
   private readonly config: IBeaconConfig;
   private readonly logger: ILogger;
@@ -45,7 +44,6 @@ export class BeaconSync extends (EventEmitter as {new (): SyncEventEmitter}) imp
   private peerCountTimer?: NodeJS.Timeout;
 
   constructor(opts: ISyncOptions, modules: ISyncModules) {
-    super();
     this.opts = opts;
     this.config = modules.config;
     this.network = modules.network;
@@ -90,7 +88,7 @@ export class BeaconSync extends (EventEmitter as {new (): SyncEventEmitter}) imp
     }
     this.mode = SyncMode.STOPPED;
     this.chain.removeListener("unknownBlockRoot", this.onUnknownBlockRoot);
-    this.regularSync.removeListener("regularSyncCompleted", this.syncCompleted);
+    this.regularSync.removeListener("syncCompleted", this.syncCompleted);
     this.stopSyncTimer();
     await this.initialSync.stop();
     await this.regularSync.stop();
@@ -153,14 +151,14 @@ export class BeaconSync extends (EventEmitter as {new (): SyncEventEmitter}) imp
     this.mode = SyncMode.REGULAR_SYNCING;
     await this.initialSync.stop();
     this.startSyncTimer(3 * this.config.params.SECONDS_PER_SLOT * 1000);
-    this.regularSync.on("regularSyncCompleted", this.syncCompleted);
+    this.regularSync.on("syncCompleted", this.syncCompleted);
     await this.gossip.start();
     await this.regularSync.start();
   }
 
-  private syncCompleted = (): void => {
+  private syncCompleted = async (): Promise<void> => {
     this.stopSyncTimer();
-    this.emit("syncCompleted");
+    await this.network.handleSyncCompleted();
   };
 
   private startSyncTimer(interval: number): void {
