+++ packages/lodestar/src/tasks/index.ts
@@ -12,9 +12,6 @@ import {ILogger} from "@chainsafe/lodestar-utils/lib/logger";
 import {IBeaconSync} from "../sync";
 import {InteropSubnetsJoiningTask} from "./tasks/interopSubnetsJoiningTask";
 import {INetwork} from "../network";
-import {DiversifyPeersBySubnetTask} from "./tasks/diversifyPeersBySubnetTask";
-import {IReputationStore} from "../sync/IReputation";
-import {CheckPeerAliveTask} from "./tasks/checkPeerAliveTask";
 
 export interface ITasksModules {
   db: IBeaconDb;
@@ -22,7 +19,6 @@ export interface ITasksModules {
   chain: IBeaconChain;
   sync: IBeaconSync;
   network: INetwork;
-  reps: IReputationStore;
 }
 
 /**
@@ -36,11 +32,8 @@ export class TasksService implements IService {
   private readonly sync: IBeaconSync;
   private readonly network: INetwork;
   private readonly logger: ILogger;
-  private readonly reps: IReputationStore;
 
   private interopSubnetsTask: InteropSubnetsJoiningTask;
-  private diversifyPeersTask: DiversifyPeersBySubnetTask;
-  private checkPeerAliveTask: CheckPeerAliveTask;
 
   public constructor(config: IBeaconConfig, modules: ITasksModules) {
     this.config = config;
@@ -49,20 +42,9 @@ export class TasksService implements IService {
     this.logger = modules.logger;
     this.sync = modules.sync;
     this.network = modules.network;
-    this.reps = modules.reps;
     this.interopSubnetsTask = new InteropSubnetsJoiningTask(this.config, {
       chain: this.chain,
       network: this.network,
-      logger: this.logger,
-    });
-    this.diversifyPeersTask = new DiversifyPeersBySubnetTask(this.config, {
-      network: this.network,
-      reps: this.reps,
-      logger: this.logger,
-    });
-    this.checkPeerAliveTask = new CheckPeerAliveTask(this.config, {
-      network: this.network,
-      reps: this.reps,
       logger: this.logger,
     });
   }
@@ -71,14 +53,13 @@ export class TasksService implements IService {
     this.chain.forkChoice.on("prune", this.handleFinalizedCheckpointChores);
     this.network.gossip.on("gossip:start", this.handleGossipStart);
     this.network.gossip.on("gossip:stop", this.handleGossipStop);
-    this.sync.on("syncCompleted", this.handleSyncCompleted);
   }
 
   public async stop(): Promise<void> {
     this.chain.forkChoice.removeListener("prune", this.handleFinalizedCheckpointChores);
     this.network.gossip.removeListener("gossip:start", this.handleGossipStart);
     this.network.gossip.removeListener("gossip:stop", this.handleGossipStop);
-    await Promise.all([this.interopSubnetsTask.stop(), this.diversifyPeersTask.stop(), this.checkPeerAliveTask.stop()]);
+    await this.interopSubnetsTask.stop();
   }
 
   private handleGossipStart = async (): Promise<void> => {
@@ -87,10 +68,6 @@ export class TasksService implements IService {
 
   private handleGossipStop = async (): Promise<void> => {
     await this.interopSubnetsTask.stop();
-  };
-
-  private handleSyncCompleted = async (): Promise<void> => {
-    await Promise.all([this.diversifyPeersTask.start(), this.checkPeerAliveTask.start()]);
   };
 
   private handleFinalizedCheckpointChores = async (finalized: BlockSummary, pruned: BlockSummary[]): Promise<void> => {
